requirements.txt
fastapi
uvicorn[standard]==0.24.0
pydantic
langchain
langchain-openai
langchain-community
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
langgraph
twilio
boto3
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
alembic==1.13.0
psycopg2-binary
redis==5.0.1
pytest-asyncio==0.21.1
reportlab==4.0.4
ddgs
